v1.0 fix

+++ src/ExpressionCalculator/Bilangan.java
@@ -57,5 +57,5 @@ public abstract class Bilangan extends Token{
      *
      * @return double
      */
-    public abstract double GetValue(); //deffered
+    public abstract int GetValue(); //deffered
 }
